Added Hazard Labels, refactored autocomplete
