+++ cfr_tool/packaging.py
@@ -5,6 +5,7 @@ from flask import (
 from . import db
 from . import clean_text as ct
 from . import code_lookup
+from . import autocomplete
 
 bp = Blueprint('packaging', __name__)
         
@@ -16,20 +17,6 @@ def check_packaging(unna, db):
         #TO DO: render packaging.html so it shows the multiple PG options for the user to select.
         render_template('packaging.html')
 
-def build_autocomplete(db):
-    db_query = '''
-    SELECT proper_shipping_name AS label, unna_code AS value
-    FROM hazmat_table
-    JOIN proper_shipping_names ON hazmat_table.row_id = proper_shipping_names.row_id
-    WHERE unna_code IS NOT NULL
-    '''
-    cursor = db.execute(db_query)
-    results = cursor.fetchall()
-
-    # Transform results into the desired format
-    autocomplete_list = [{'label': row[0].strip().replace('\n',''), 'value': row[1]} for row in results]
-    return autocomplete_list
-
 @bp.route('/help', methods=['GET'])
 def help():
     return render_template('help.html')
@@ -40,7 +27,7 @@ def help():
 
 @bp.route('/',  methods=('GET', 'POST'))
 def packaging():
-    autocomplete_data = build_autocomplete(db.get_db())
+    autocomplete_data = autocomplete.build_autocomplete(db.get_db())
     if request.method == 'POST':
         un_id = request.form['un_id']
         try:
@@ -72,4 +59,3 @@ def packaging():
             return render_template('packaging.html', autocomplete_data=autocomplete_data)
         flash(error)
     return render_template('packaging.html', autocomplete_data=autocomplete_data)
-
