+++ start.bat
@@ -3,3 +3,4 @@ set FLASK_APP=cfr_tool
 set FLASK_ENV=development
 flask init-db
 flask run
+pause
